// server/src/enums/DeviceType.ts
enum DeviceType {
	VaasanSahkoIntegration = "VaasanSahkoIntegration",
	MatterIntegration = "MatterIntegration",
	WizIntegration = "WizIntegration",
	ShellyIntegration = "ShellyIntegration",
	ToshibaAcIntegration = "ToshibaAcIntegration",
	RestApiIntegration = "RestApiIntegration"
}

export default DeviceType;